move date formatting utils to common/i18n
