+++ ui/common/tests/timeago.test.ts
@@ -12,11 +12,11 @@ describe('test formatter', () => {
   ])('lang code formatting', async(lang, expected) => {
     document.documentElement.lang = lang;
 
-    const displayLocale = await import('../src/timeago').then(m => m.displayLocale);
+    const displayLocale = await import('../src/i18n').then(m => m.displayLocale);
     expect(displayLocale).toBe(lang);
 
-    const formatter = await import('../src/timeago').then(m => m.formatter);
-    expect(formatter()(new Date(2024, 0, 1, 17, 0, 0))).toBe(expected);
+    const formatter = await import('../src/i18n').then(m => m.commonDateFormat);
+    expect(formatter(new Date(2024, 0, 1, 17, 0, 0))).toBe(expected);
   });
 
   test.each([
@@ -39,10 +39,10 @@ describe('test formatter', () => {
   ])('arabic lang code uses the gregorian calendar', async lang => {
     document.documentElement.lang = lang;
 
-    const displayLocale = await import('../src/timeago').then(m => m.displayLocale);
+    const displayLocale = await import('../src/i18n').then(m => m.displayLocale);
     expect(displayLocale).toBe('ar-ly');
 
-    const formatter = await import('../src/timeago').then(m => m.formatter);
-    expect(formatter()(new Date(2024, 0, 1, 17, 0, 0))).toBe('1 يناير 2024، 5:00 م');
+    const formatter = await import('../src/i18n').then(m => m.commonDateFormat);
+    expect(formatter(new Date(2024, 0, 1, 17, 0, 0))).toBe('1 يناير 2024، 5:00 م');
   });
 });
